+++ lib/core/routes.dart
@@ -1,3 +1,5 @@
+import 'package:app/features/home/home.screen.dart';
+import 'package:app/features/quran/quran.screen.dart';
 import 'package:flutter/material.dart';
 
 import '../features/auth/login.screen.dart';
@@ -7,5 +9,7 @@ class AppRoutes {
   static Map<String, WidgetBuilder> routes = {
     LoginScreen.routeName: (context) => const LoginScreen(),
     NavigationScreen.routeName: (context) => const NavigationScreen(),
+    HomeScreen.routeName: (context) => const HomeScreen(),
+    QuranScreen.routeName: (context) => const QuranScreen(),
   };
 }