+++ lib/features/home/widgets/features.card.dart
@@ -8,30 +8,35 @@ class FeaturesCard extends StatelessWidget {
       required this.width,
       required this.height,
       required this.title,
-      required this.backgroundImage});
+      required this.backgroundImage,
+      this.routeName});
   final double width;
   final double height;
   final String title;
   final String backgroundImage;
+  final String? routeName;
 
   @override
   Widget build(BuildContext context) {
-    return Container(
-      width: width,
-      height: height,
-      padding: EdgeInsets.all(kSpacingX2),
-      decoration: BoxDecoration(
-        image: DecorationImage(
-          image: AssetImage(backgroundImage),
-          fit: BoxFit.cover,
-          colorFilter: const ColorFilter.mode(Colors.grey, BlendMode.darken),
+    return InkWell(
+      onTap: () => context.pushNamed(routeName ?? "/"),
+      child: Container(
+        width: width,
+        height: height,
+        padding: EdgeInsets.all(kSpacingX2),
+        decoration: BoxDecoration(
+          image: DecorationImage(
+            image: AssetImage(backgroundImage),
+            fit: BoxFit.cover,
+            colorFilter: const ColorFilter.mode(Colors.grey, BlendMode.darken),
+          ),
+          borderRadius: BorderRadius.circular(kSpacingX2),
         ),
-        borderRadius: BorderRadius.circular(kSpacingX2),
-      ),
-      child: Text(
-        title.translate(context),
-        maxLines: 3,
-        style: context.textTheme.headlineMedium,
+        child: Text(
+          title.translate(context),
+          maxLines: 3,
+          style: context.textTheme.headlineMedium,
+        ),
       ),
     );
   }