+++ CHANGELOG.md
@@ -4,6 +4,38 @@ All notable changes to this project will be documented in this file. The format 
 
 ## [Unreleased]
 
+## [0.2.0] - 2024-01-25
+
+- **Release Date**: 2024-01-25
+- **Commit Message**: `Added islamic screen and quran screen`
+
+### Added
+
+- **Quran**
+  - Description: Added Screen to show the progress of the user in the quran and the list of surats that the user can read or learn about.
+  - Author: @bloopdex
+
+### Fixed
+
+- **Routes**
+  - Description: Fixed the routes of the application to make it more readable and easier to maintain.
+  - Author: @bloopdex
+- **Routes**
+  - Description: Fixed the routes of the application to make it more readable and easier to maintain.
+  - Author: @bloopdex
+
+### Changed
+
+- **User Model**
+  - Description: Changed the user model to add percentage of the progress in the quran.
+  - Author: @bloopdex
+- **Islamic Screen Template Scaffold**
+  - Description: Changed the template of the islamic screen to add a default appbar that has a back button.
+  - Author: @bloopdex
+- **Custom Button**
+  - Description: Changed the custom button to add width parameters.
+  - Author: @bloopdex
+
 ## [0.1.0] - 2024-01-25
 
 - **Release Date**: 2024-01-25